Restruct db and controller interfaces
Need to easily migrate to other db, or test with mock db's. Make db
as an IModel interface, which can return the last unique id which
would be converted to base62 and saved to the db during encode.

The GetLastId should 'always' return a unique id. Depend on db implementation,
for example NoSql db's can use random numbers, but even big range of random numbers,
collision does and will occur, so it should be the responsblity of the Model
which implement the GetLastId to be smart about it. For now we resort to using
the SQL db with the table id increments as last id.

Added fail safe retry mechanism just in case if the LastId is not as unique as
expected.